+++ anikanov.vladislav/T1/square.cpp
@@ -44,7 +44,7 @@ void Square::move(const float x, const float y)
   cPoint = point_t(x, y);
 }
 
-void Square::scale(const float k, const point_t center)
+void Square::scale(const double k, const point_t center)
 {
   if (k < 0) {
     throw std::logic_error("Invalid scale argument");
